pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["Alex Angelico <alex.angelico@live.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
